Add benchmark, improve mapper, and improve age (#25)
* add compare function to replace common TREE_EQ

* added a basic benchmark for coordinate mapping

* fix shape and upgrade benchmark

* add more benchmarks and move to ade

* implement reduce_id mapper

* implement bijective check in iCoordMap

* add default value and descriptions to api

* remove executable flag for dot files and update docs

* add test for reduce_1d_map and update comments

+++ dbg/tensor.hpp
@@ -58,7 +58,7 @@ private:
 		else
 		{
 			auto it = shape.begin();
-			size_t before = std::accumulate(it, it + rank, 1,
+			size_t before = std::accumulate(it, it + rank, (size_t) 1,
 				std::multiplies<size_t>());
 			print_helper(out, shape, arr, rank - 1);
 			for (uint16_t i = 1; i < n; ++i)
